Edit code styles to ensure coding standard
