+++ build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'java'
     id 'application'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
+    id 'checkstyle'
 }
 
 repositories {
@@ -31,6 +32,10 @@ dependencies {
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'linux'
 }
 
+checkstyle {
+    toolVersion = '10.2'
+}
+
 test {
     useJUnitPlatform()
 
@@ -46,7 +51,7 @@ test {
 }
 
 application {
-    mainClass.set("Vic.Vic")
+    mainClass.set("vic.Vic")
 }
 
 shadowJar {
