--- src/main/java/Vic/exceptions/TaskOutOfBoundsException.java
@@ -1,10 +1,0 @@
-package Vic.exceptions;
-
-/**
- * This exception is thrown when an invalid task ID is provided.
- */
-public class TaskOutOfBoundsException extends Exception {
-    public TaskOutOfBoundsException() {
-        super("The task id provided is invalid! (⚆_⚆)");
-    }
-}
