--- src/main/java/Vic/exceptions/VicException.java
@@ -1,10 +1,0 @@
-package Vic.exceptions;
-
-/**
- * This exception provides a generic error message
- */
-public class VicException extends Exception {
-    public VicException() {
-        super("Something went wrong! Please contact the developer for help! (╥╯ᗝ╰╥)");
-    }
-}
