--- src/main/java/Vic/tasks/Task.java
@@ -1,66 +1,0 @@
-package Vic.tasks;
-
-/**
- * Represents a task with a description and completion status.
- */
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    /**
-     * Returns a string representation of the task, including its completion status.
-     *
-     * @return A string representing the task's status and description.
-     */
-    @Override
-    public String toString() {
-        return "[" + getStatusIcon() + "] " + description;
-    }
-
-    /**
-     * Returns the description of the task.
-     *
-     * @return The description of the task.
-     */
-    public String getDescription() {
-        return description;
-    }
-
-    /**
-     * Returns the status of the task as an icon.
-     * "X" indicates the task is completed, while " " indicates it is incomplete.
-     *
-     * @return A string representation of the task's status icon.
-     */
-    public String getStatusIcon() {
-        return (isDone ? "X" : " ");
-    }
-
-    /**
-     * Returns the completion status of the task.
-     *
-     * @return True if the task is marked as done, false otherwise.
-     */
-    public boolean getStatus() {
-        return isDone;
-    }
-
-    /**
-     * Marks the task as completed.
-     */
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    /**
-     * Marks the task as incomplete.
-     */
-    public void markAsUndone() {
-        isDone = false;
-    }
-}
