+++ src/main/java/vic/actions/ByeAction.java
@@ -1,0 +1,27 @@
+package vic.actions;
+
+import vic.storage.Storage;
+import vic.tasks.TaskList;
+
+/**
+ * Handles marking tasks as done or undone
+ */
+public class ByeAction extends Action {
+
+    /**
+     * Constructor for class
+     */
+    public ByeAction(Storage storage, TaskList taskList, String action) {
+        super(storage, taskList, action);
+    }
+
+    /**
+     * Ends the action
+     *
+     * @return true to exit the application
+     */
+    @Override
+    public boolean execute() {
+        return true;
+    }
+}
